Remove SUNXI_KERNEL option and related ifdef's
Utility nrf24_test needs to be tuned for a specific Linux kernel. Here is the list of
areas where different kernels may need different settings:
- gpio numbers and names
- spi mode (sunxi spi driver needs LOOP parameter for proper 'full-duplex' operation)
- proper selection of spidev.h header
  -- using default spidev.h from linux-headers is not recommented: this part of kernel UAPI is quite unstable
  -- it is more safe to use spidev.h from the kernel in use

SUNXI_KERNEL was introduced to distinguish between two different linux kernels: sunxi-3.4
and upstream-4.0. It was a temporary solution for board bringup using two different
kernels. Now it is time to start work on a more flexible solution. This commit implements
the following steps on this way:
- remove SUNXI_KERNEL and related ifdefs (for now keep settings only for sunxi-3.4 kernel)
- introduce KERNEL_DIR cmake parameter which points at the source dir of the kernel in use
  -- this option needs to be set, otherwise compiler will not find a way to spidev.h
  -- for configuration use the following cmake option: -DKERNEL_DIR:PATH=/path/to/kernel/dir

TODO: introduce config file with spi and gpio settings

Signed-off-by: Sergey Matyukevich <geomatsi@gmail.com>