Add support for multi-byte SPI transfers
This patch adds support for multi-byte SPI transfers and enables driver for
Orange-Pi One board. Here is a summary of changes:

- add multi-byte SPI transfers support to spidev

- setup Orange-Pi One driver featuring the following configuration
  -- multi-byte SPI transfers
  -- hardware SPI chip select line: no need for GPIO line
  -- no support for single-byte SPI transfers

- cleanup: remove set_spi_speed, it is not supported by updated libnrf24

Notes.
1. Problem with single-byte transfers on Orange-Pi One.
Orange-Pi driver completely drops single-byte SPI transfers to nRF24XXX chip
since they does not work properly on that board. Here is a brief overview of the
problem with single-byte xfers. For some reason sun8i SPI controller toggles
additional CLK pulse on SPI transfer setup. When GPIO CSN is used, then it is
set before the SPI transfer is configured and started. nRF24 SPI controller is
waiting for CLK pulses after CSN is down. So it is confused by additional CLK
pulse: it considers it as a first valid CLK pulse and starts capturing bits. As
a result, the communication is broken.

On the other hand, when h/w controlled CSN line on Orange-Pi One is used, then
it is pulled down by sun8i SPI hardware right _after_ the confusing extra CLK
pulse. As a result, this extra CLK is ignored by nRF24 (since CSN is still high at
that time).

2. Selection between single- and multi- byte SPI transfers
It is possible to implement both methods and pass them to rf24_init. Actual
choice is done by linking to a proper libnrf24 flavor:
- if app is linked with libnrf24.a, then single-byte method will be used
- if app is linked with libnrf24_mb.a, then multi-byte method will be used

So if only one method is implemented and the other one is set to NULL, then make
sure that app is linked against the proper library. Otherwise app will crash
since currently there is no run-time check in libnrf24.

Signed-off-by: Sergey Matyukevich <geomatsi@gmail.com>